% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dentist.R
\name{dent_likelihood}
\alias{dent_likelihood}
\title{Dents the likelihood surface
This takes any values that are better (lower) than the desired negative log likelihood and reflects them across the best_neglnL + delta line, "denting" the likelihood surface.}
\usage{
dent_likelihood(neglnL, best_neglnL, delta = 2)
}
\arguments{
\item{neglnL}{The original negative log likelihood}

\item{best_neglnL}{The negative log likelihood at the optimum; other values will be greater than this.}

\item{delta}{How far from the optimal negative log likelihood to focus samples}
}
\value{
The transformed negative log likelihood
}
\description{
Dents the likelihood surface
This takes any values that are better (lower) than the desired negative log likelihood and reflects them across the best_neglnL + delta line, "denting" the likelihood surface.
}
