% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dentist.R
\name{dent_propose}
\alias{dent_propose}
\title{Propose new values
This proposes new values using a normal distribution centered on the original parameter values, with desired standard deviation. If any proposed values are outside the bounds, it will propose again.}
\usage{
dent_propose(old_params, lower_bound = -Inf, upper_bound = Inf, sd = 1)
}
\arguments{
\item{old_params}{The original parameter values}

\item{lower_bound}{Minimum parameter values to try. One for all or a vector of the length of par.}

\item{upper_bound}{Maximum parameter values to try. One for all or a vector of the length of par.}

\item{sd}{Standard deviation to use for the proposals. One for all or a vector of the length of par.}
}
\value{
A vector of the new parameter values
}
\description{
Propose new values
This proposes new values using a normal distribution centered on the original parameter values, with desired standard deviation. If any proposed values are outside the bounds, it will propose again.
}
