\name{primates}
\docType{data}
\alias{tree}
\alias{trait}
\alias{rayDISC.example.tree}
\alias{rayDISC.example.data}
\title{An example dataset}
\description{
The primate dataset from BayesTraits which contains a single tree with branch lengths scaled to time, and a trait file that contains two binary characters. 
}
\format{a tree of class \dQuote{phylo} and a data frame with 3 columns and 60 rows}
\references{
Pagel, M., and A. Meade. 2006. Bayesian analysis of correlated evolution of discrete characters by reversible-jump Markov chain Monte Carlo. American Naturalist 167:808-825.
}

\keyword{datasets}