\name{primetes}
\docType{data}
\alias{tree}
\alias{trait}
\alias{rayDISC.example.tree}
\alias{rayDISC.example.data}
\alias{.Random.seed}
\title{An example dataset}
\description{
An example dataset containing a 64-tip birth-death tree with internal node labels denoting two selective regimes, and a trait file in the proper format: 1) Genus_species, 2) current selective regime, 3) continuous trait data. 
}
\format{a tree of class \dQuote{phylo} and a data frame with 3 columns and 64 rows}
\keyword{datasets}