% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dentist.R
\name{summary.dentist}
\alias{summary.dentist}
\title{Summarize dentist
Display summary of output from dent_walk}
\usage{
\method{summary}{dentist}(object, ...)
}
\arguments{
\item{object}{An object of class dentist}

\item{...}{Other arguments (not used)}
}
\description{
Summarize dentist
Display summary of output from dent_walk
}
